+++ kai/ukernels/matmul/pack/kai_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme.c
@@ -3,7 +3,11 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 //
+#if (!defined(__aarch64__) || !defined(__ARM_FEATURE_SVE2)) && !defined(_M_ARM64)
+#error This file must be compiled for AArch64, FEAT_SVE2.
+#else  // Architectural features check.
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "kai/kai_common.h"
@@ -11,16 +15,13 @@
 #include "kai/ukernels/matmul/kai_matmul_pack_lhs_types.h"
 
 enum {
-    RHS_ELEM_BYTES = 1,
-
+    OUTPUT_ELEM_BYTES = 1,
     MR_VSCALE = 4,
     KR = 4,
-
     MAX_MR = MR_VSCALE * KAI_VSCALE_MAX,
 };
 
-void kai_kernel_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme(
-    size_t height, size_t width, const void* in, size_t row_offset, void* out);
+void kai_matmul_pack_rows_x8p4vsx4_x8_sme(size_t height, size_t width, const void* in, void* out);
 
 static size_t get_mr(void) {
     return MR_VSCALE * kai_get_sme_vscale();
@@ -30,20 +31,12 @@ static size_t div_ceil(size_t a, size_t b) {
     return (a + b - 1) / b;
 }
 
-static size_t get_m_step(void) {
-    return get_mr();
-}
-
-static size_t get_k_step(void) {
-    return 0;
-}
-
 static struct kai_matmul_pack_lhs_uker_dim_args get_step(const struct kai_matmul_pack_lhs_uker_config* config) {
     KAI_UNUSED(config);
 
     const struct kai_matmul_pack_lhs_uker_dim_args step = {
-        .m = get_m_step(),
-        .k = get_k_step(),
+        .m = get_mr(),
+        .k = 0,
     };
 
     return step;
@@ -54,7 +47,7 @@ static struct kai_matmul_pack_lhs_uker_lhs_stride_args get_lhs_stride(
     KAI_UNUSED(config);
 
     const struct kai_matmul_pack_lhs_uker_lhs_stride_args stride = {
-        .m = shape->k * RHS_ELEM_BYTES,
+        .m = shape->k * OUTPUT_ELEM_BYTES,
     };
 
     return stride;
@@ -64,10 +57,10 @@ static size_t get_lhs_offset(
     const struct kai_matmul_pack_lhs_uker_config* config, const struct kai_matmul_pack_lhs_uker_lhs_dim_args* index,
     const struct kai_matmul_pack_lhs_uker_lhs_stride_args* stride) {
     KAI_UNUSED(config);
-    KAI_ASSUME(index->m % get_m_step() == 0);
+    KAI_ASSUME(index->m % get_mr() == 0);
     KAI_ASSUME(index->k == 0);
 
-    return index->m * stride->m + index->k * RHS_ELEM_BYTES;
+    return index->m * stride->m + index->k * OUTPUT_ELEM_BYTES;
 }
 
 static struct kai_matmul_pack_lhs_uker_lhs_packed_stride_args get_lhs_packed_stride(
@@ -77,7 +70,7 @@ static struct kai_matmul_pack_lhs_uker_lhs_packed_stride_args get_lhs_packed_str
 
     const size_t mr = get_mr();
     const struct kai_matmul_pack_lhs_uker_lhs_packed_stride_args stride = {
-        .m = mr * kai_roundup(shape->k, KR) * RHS_ELEM_BYTES,
+        .m = mr * kai_roundup(shape->k, KR) * OUTPUT_ELEM_BYTES,
     };
 
     return stride;
@@ -88,11 +81,11 @@ static size_t get_lhs_packed_offset(
     const struct kai_matmul_pack_lhs_uker_lhs_packed_dim_args* index,
     const struct kai_matmul_pack_lhs_uker_lhs_packed_stride_args* stride) {
     KAI_UNUSED(config);
-    KAI_ASSUME(index->m % get_m_step() == 0);
+    KAI_ASSUME(index->m % get_mr() == 0);
     KAI_ASSUME(index->k == 0);
 
     const size_t mr = get_mr();
-    return index->m / mr * stride->m + index->k * mr * RHS_ELEM_BYTES;
+    return index->m / mr * stride->m + index->k * mr * OUTPUT_ELEM_BYTES;
 }
 
 static size_t get_lhs_packed_size(
@@ -104,36 +97,35 @@ static size_t get_lhs_packed_size(
     const size_t mr = get_mr();
     return div_ceil(shape->m, mr) * stride->m;
 }
-
 static void run(
     const struct kai_matmul_pack_lhs_uker_config* config, const struct kai_matmul_pack_lhs_uker_args* args) {
     KAI_UNUSED(config);
 
     const size_t mr = get_mr();
 
-    const size_t height = args->shape.m;
+    const size_t m = args->shape.m;
     const size_t width = args->shape.k;
 
     const uint8_t* lhs_ptr = args->operand.lhs.ptr;
     uint8_t* lhs_packed_ptr = args->operand.lhs_packed.ptr;
 
-    const uint8_t* src_ptrs[MAX_MR];
-
+    const uint8_t* in[MAX_MR];
     kai_commit_za();
 
-    for (size_t start_row = 0; start_row < height; start_row += mr) {
-        const size_t block_height = KAI_MIN(height - start_row, mr);
+    for (size_t start_row = 0; start_row < m; start_row += mr) {
+        const size_t height = KAI_MIN(m - start_row, mr);
+        KAI_UNUSED(start_row);
 
-        uint8_t* dst = lhs_packed_ptr;
+        void* out = lhs_packed_ptr;
         lhs_packed_ptr += args->operand.lhs_packed.stride.m;
 
-        for (size_t row = 0; row < block_height; ++row) {
-            src_ptrs[row] = lhs_ptr + row * args->operand.lhs.stride.m;
+        for (size_t row = 0; row < height; ++row) {
+            in[row] = lhs_ptr + row * args->operand.lhs.stride.m;
         }
         lhs_ptr += mr * args->operand.lhs.stride.m;
 
         // NOLINTNEXTLINE(bugprone-multi-level-implicit-pointer-conversion)
-        kai_kernel_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme(block_height, width, src_ptrs, 0, dst);
+        kai_matmul_pack_rows_x8p4vsx4_x8_sme(height, width, in, out);
     }
 }
 
@@ -153,3 +145,5 @@ struct kai_matmul_pack_lhs_uker_api kai_matmul_pack_lhs_mxk_x8p4vsx4_x8_sme(void
 
     return api;
 }
+
+#endif  // Architectural features check.