+++ test/common/sme.hpp
@@ -21,9 +21,7 @@ uint64_t get_sme_vector_length() {
     return get_sme_vector_length<sizeof(T)>();
 }
 
-/// Gets the SME vector scale.
-inline uint64_t get_sme_vscale() {
-    return get_sme_vector_length<uint8_t>() / 16;
-}
+/// Gets the SME vector scale
+uint64_t get_sme_vector_scale();
 
 }  // namespace kai::test